packaging/containernetworking-cni.spec - COPR
I would like to start building RPMs in my COPY repo
for CNI as part of "Upstream_CRIO_Family". We would build
new packages on each merge based on a webhook between
COPR and github.

https://copr.fedorainfracloud.org/coprs/baude/Upstream_CRIO_Family/

This does not replace the Fedora RPMs but does allow us to work
with the most recent CNI code in an easier fashion.

Signed-off-by: baude <bbaude@redhat.com>