macvlan: Create macvtap at a random index
There is a limitation in the linux kernel that prevents a macvtap link
from getting the correct link index when created in a network namespace.

We need to pick a random non conflicting index and try to
create a link. If that fails, we need to try with another.
This is to ensure the device node created for the macvtap does not
conflict.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>