host-local: still allocate empty file to avoid IP leak
It may happen when host-local creates an empty file successfully, but fails to write
string into it and further fails to remove it. Many empty file will be created if
runtime retries to create a new container until all free IPs are leaked in the range set.
The root cause is that host-local cannot handle this kind of error when performing a
rollback. We should check if IP file is empty, this prevents exhausting all free IPs and
helps to recover leaked IP files.

Fix #498

Signed-off-by: chenyaqi01 <chenyaqi01@baidu.com>