host-device: fix host device delete if it was renamed
When a device previously passed to the container has to be moved back to
the host, netlink.LinkByName() is used to find it in the container. This
does not work if the device was renamed in the container.

Fix this by storing the initial interface name in the alias, in addition
to the host interface name. To find the device, we now iterate the link
list until we find one whose alias corresponds.

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>