bridge, spoofcheck: Apply nftable rules in the pod netns
The spoofcheck nftables rules have been applied on the node root network
namespace for ingress traffic that arrives from the pod netns.

When there are a large amount of rules visible, the user-space
processing of adding and removing rules becomes slow. Field results
showed the `nft` commands taking minutes to complete.

In addition, the removal of the rules is dependent on the execution of
the DEL command, which is not always assured to occur. Therefore, the
current solution where the rules are set on the node root network
namespace may cause nftables configuration leaks.

This change suggests to place the rules in the pod network namespace,
solving both issues.

Signed-off-by: Edward Haas <edwardh@redhat.com>