spoofcheck: Add nftables rules to the IP stack postroute hook
When the nftables rules are used in the pod netns, it is not enough to
use the bridge family postrouting hook.
For setups where the egress traffic is not passing through a bridge (in
the pod netns), there is a need to also cover the IP stack postrouting
hook.

Tables, chains and rules are set accordingly.

Signed-off-by: Edward Haas <edwardh@redhat.com>