bridge: Disallow VLAN 1 when connecting to a VLAN
The bridge CNI allows users to select a VLAN to which a Pod should be
connected:

  {
    "cniVersion": "0.3.1",
    "name": "mynet",
    "type": "bridge",
    "bridge": "mynet0",
    "vlan": 100
  }

The expected behavior in such a case is that the Pod is isolated within
the requested VLAN.

That however is not the case. Today, the underlying OS applies the
default VLAN 1 to the bridge itself and to every veth connected to it.
Due to that, the Pod connected to a VLAN is still receiving all the
native traffic of the bridge.

That can be reproduced by:
* Using the configuration listed above,
* setting IP 11.11.11.1/24 on the bridge,
* setting IP 11.11.11.10/24 in the Pod,
* broadcasting ARP on the bridge: arping -A -I mynet0 11.11.11.1
* listening to ARP in the Pod: tcpdump -lni net1 arp

With this suggested fix, the implicitly added VLAN 1 is removed in case
another VLAN ID is provided.

Fixes: https://github.com/containernetworking/plugins/issues/667

Signed-off-by: Petr Horacek <hrck@protonmail.com>