sbr: Add outbound interface rule in addition to source rule
Adding the outbound interface rule allows applications to bind to an interface name
rather than only the interface IP. This allows applications in a multus environment
to be configured with the interface name, which can be configured, rather than the
interface IP address, which is not known in advance.

The outbound interface rule is on;y added if the interface is configured with 1 IP
address, because when there are multiple, only one will be selected, depending on the
rule order, which is non-deterministic.

Signed-off-by: Neil Cook <neil.cook@noware.co.uk>