Update bridge.go
Port Isolation of an interface in the Linux bridge is only supported by a subset of interfaces. This code change checks that the interface is of type veth, vnet or tap before attempting to isolate the interface to the bridge. 

The existing code would try to isolate any interface type with no checking and would easily cause an error.